backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
supabase==2.13.0
python-dotenv==1.1.0
openai==1.78.1
pydantic==2.11.3
